WCM-285: move IConverter and create converter module
filesystem and postgresql connector will need them
